Move main.swift to `testSample`
